requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.2
boto3==1.42.3
botocore==1.42.3
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.1.2
django-allauth==65.12.1
django-cors-headers==4.9.0
django-storages==1.14.6
django-tailwind==4.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kombu==5.5.4
multidict==6.6.4
packaging==25.0
pillow==11.0.0
prompt_toolkit==3.0.52
propcache==0.3.2
# psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
qrcode==8.0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.1
stripe==12.5.1
twilio==9.8.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
yarl==1.20.1
